Make Expression and Statement concepts rather than base classes.
Previously, to define a new type of expression, you would:

  * Derive from Expression.
  * Define a constructor, since aggregate initialization doesn't play
    nice with base classes.
  * Re-declare Print(), since it is overridden.

Now, you would:

  * Define a plain struct.
  * Include anything you need, but at least include a
    `Location location` field.
  * To mark the struct as a type of expression, add an overload of
    `ExpressionVisitor::operator()` which takes that type.

The new approach avoids the boilerplate of declaring and defining
a trivial constructor, and declaring Print(). It also allows for
arbitrary `ExpressionVisitor` types, so it won't be necessary to extend
the interface for every single type in order to support later stages of
compilation.