Fix function parameters and comparison operators.
Function parameters had the wrong offsets because I wrote the code to
generate those offsets before I decided to add an implicit parameter for
the return slot.

Binary operators were broken in many cases because I used `xor` after
the `test` or `cmp` but before the `set??`, so the flags were being
overwritten.