Make the timing script spawn background work.
Running dd in the background forces the CPU to clock up. Without this,
the solutions with very short execution times might be penalised due to
the CPU running underclocked. The visible effect of this is much higher
variance in execution times without any background work running.

The README numbers are updated, this time using my desktop rather than
my laptop, so they are all a bit lower.